Create a new library project using Cargo
The main goal of this project is to make it harder to manually implement
futures that return `Poll::Pending` without having correctly scheduled
the future to be polled using the context's waker. In order to do this,
a new type will be created to be used as a token. This token type can
only be created either using an `unsafe` call, forcing the developer to
mark its creation as an area that requires attention. They can then
check that the pieces of code where a poll needs to scheduled is
correctly implemented. If a future is implemented by composing other
futures, the idea is that no `unsafe` code will be needed, since the
type system can help the developer to return `SafePoll::Pending` using
any one of the tokens obtained from polling the inner futures.