Create `WakeupRegisteredToken` zero sized type
This type is used to mark that the current task was successfully
registered for a wakeup so that it will be polled again later. The idea
is to make it easier to find situations where a `Future` needs to
manually register a wakeup. Because the token type has an `unsafe`
constructor, the developer only needs to check the `unsafe` blocks to
verify that manually implemented `Future`s are correctly registering
themselves for wakeups.