Add some initial crate documentation
Describe the goals of the crate and provide some initial examples. The
last code snippet shows how the crate is planned to be used, but since
it depends on things that haven't been implemented yet, it's execution
as a documentation test is currently ignored. That should change once
the crate provides the necessary parts for it to run successfully.